+++ backend/app/lib/agent_tools/__init__.py
@@ -1,8 +1,6 @@
-from .travel_tools import (
-    GoogleFlightsTool,
-    GoogleHotelsTool,
-    TavilySearchTool,
-)
+from .google_flight_tool import GoogleFlightsTool
+from .google_hotel_tool import GoogleHotelsTool
+from .tavily_search_tool import TavilySearchTool
 
 __all__ = [
     "GoogleFlightsTool",
